+++ cogs/notification.py
@@ -13,7 +13,7 @@ from src.log import setup_logger
 from src.notification.account_tracker import AccountTracker
 from src.permission import ADMINISTRATOR
 from src.db_function.readonly_db import connect_readonly
-from src.utils import get_accounts, get_lock, get_utcnow
+from src.utils import get_accounts, get_lock, get_utcnow, validate_and_normalize_language
 from src.presence_updater import update_presence
 
 log = setup_logger(__name__)
@@ -80,7 +80,7 @@ class Notification(Cog_Extension):
                         if match_user is None:
                             async with lock:
                                 await db.execute('BEGIN')
-                                await cursor.execute('INSERT INTO user (id, username, lastest_tweet, client_used) VALUES (?, ?, ?, ?)', (str(new_user.id), new_user.username, get_utcnow(), account_used))
+                                await cursor.execute('INSERT INTO user (id, username, latest_tweet, client_used) VALUES (?, ?, ?, ?)', (str(new_user.id), new_user.username, get_utcnow(), account_used))
                                 await cursor.execute('INSERT OR IGNORE INTO channel VALUES (?, ?)', (str(channel.id), server_id))
                                 await cursor.execute('INSERT INTO notification (user_id, channel_id, role_id, enable_type, enable_media_type) VALUES (?, ?, ?, ?, ?)', (str(new_user.id), str(channel.id), roleID, enable_type, media_type))
                                 await db.commit()
@@ -142,9 +142,9 @@ class Notification(Cog_Extension):
         if match_user is None or match_user['enabled'] == 0:
             await self.account_tracker.addTask(new_user.username, account_used)
             await update_presence(self.bot)
-            await itn.followup.send(f'successfully add notifier of {new_user.username} under {account_used}!', ephemeral=True)
+            await itn.followup.send(f'successfully added notifier for new user {new_user.username} to {account_used}!', ephemeral=True)
         else:
-            await itn.followup.send(f'{match_user["username"]} already exists under {match_user["client_used"]}. Using the same account to deliver notifications', ephemeral=True)
+            await itn.followup.send(f'successfully added/updated notifier for {match_user["username"]} (via {match_user["client_used"]})', ephemeral=True)
 
     @remove_group.command(name='notifier')
     @app_commands.rename(channel_id='channel')
@@ -255,6 +255,48 @@ class Notification(Cog_Extension):
                 else:
                     await itn.response.send_message(f'can\'t find notifier {username} in {channel.mention}!', ephemeral=True)
                     
+    @customize_group.command(name='translation')
+    async def customize_translation(self, itn: discord.Interaction, username: str, language: str = None):
+        """Set customized translation language for a tracked account.
+
+        Parameters
+        -----------
+        username: str
+            The username of the tracked account you want to set customized translation language.
+        language: str
+            The language code you want to translate to (e.g. en, ja). Leave it empty to use default.
+        """
+        if configs['embed']['type'] != 'proxy' or not configs['embed']['proxy']['auto_translation']['enabled']:
+            await itn.response.send_message('translation is not enabled in configs!', ephemeral=True)
+            return
+        
+        lang_code = validate_and_normalize_language(language)
+        if language and lang_code is None:
+            await itn.response.send_message('invalid language code!', ephemeral=True)
+            return
+
+        await itn.response.defer(ephemeral=True)
+
+        async with aiosqlite.connect(os.path.join(os.getenv('DATA_PATH'), 'tracked_accounts.db')) as db:
+            await db.execute('PRAGMA synchronous = OFF')
+            await db.execute('PRAGMA count_changes = OFF')
+
+            db.row_factory = aiosqlite.Row
+            async with db.cursor() as cursor:
+                await cursor.execute('SELECT user.id FROM user JOIN notification ON user.id = notification.user_id JOIN channel ON notification.channel_id = channel.id WHERE username = ? COLLATE NOCASE AND channel.server_id = ? AND notification.enabled = 1', (username, str(itn.guild_id)))
+                match_user = await cursor.fetchone()
+                if match_user is not None:
+                    async with lock:
+                        await cursor.execute('INSERT OR REPLACE INTO server_user_config (server_id, user_id, translate) VALUES (?, ?, ?)', (str(itn.guild_id), match_user['id'], lang_code))
+                        await db.commit()
+                    
+                    if language is None:
+                        await itn.followup.send(f'successfully restored {username}\'s translation setting to default ({configs["embed"]["proxy"]["auto_translation"]["default_language"]})', ephemeral=True)
+                    else:
+                        await itn.followup.send(f'successfully set {username}\'s translation language to {lang_code}!', ephemeral=True)
+                else:
+                    await itn.followup.send(f'can\'t find twitter user {username} in this server!', ephemeral=True)
+
     @r_notifier.autocomplete('channel_id')
     async def get_channels_for_r_notifier(self, itn: discord.Interaction, input_channel: str) -> list[app_commands.Choice[str]]:        
         return await fetch_tracked_channels(itn, input_channel, include_unknown=True)
@@ -274,6 +316,15 @@ class Notification(Cog_Extension):
             db.row_factory = aiosqlite.Row
             async with db.cursor() as cursor:
                 await cursor.execute('SELECT user.username FROM user JOIN notification ON user.id = notification.user_id WHERE notification.channel_id = ? AND notification.enabled = 1', (selected_channel_id,))
+                users = [row['username'] async for row in cursor]
+                return [app_commands.Choice(name=row, value=row) for row in users if username.lower() in row.lower()]
+
+    @customize_translation.autocomplete('username')
+    async def get_guild_enabled_users(self, itn: discord.Interaction, username: str) -> list[app_commands.Choice[str]]:
+        async with connect_readonly(os.path.join(os.getenv('DATA_PATH'), 'tracked_accounts.db')) as db:
+            db.row_factory = aiosqlite.Row
+            async with db.cursor() as cursor:
+                await cursor.execute('SELECT user.username FROM user JOIN notification ON user.id = notification.user_id JOIN channel ON notification.channel_id = channel.id WHERE channel.server_id = ? AND notification.enabled = 1', (str(itn.guild_id),))
                 users = [row['username'] async for row in cursor]
                 return [app_commands.Choice(name=row, value=row) for row in users if username.lower() in row.lower()]